leverage utils
