+++ src/worker/vm/upgrades/historical-export/utils.ts
@@ -118,46 +118,6 @@ export const fetchEventsForInterval = async (
     }
 }
 
-// This assumes the value stored at `key` can be cast to a Postgres numeric type
-export const postgresIncrement = async (
-    db: DB,
-    pluginConfigId: PluginConfig['id'],
-    key: string,
-    incrementBy = 1
-): Promise<number> => {
-    const incrementResult = await db.postgresQuery(
-        `
-        INSERT INTO posthog_pluginstorage (plugin_config_id, key, value)
-        VALUES ($1, $2, $3)
-        ON CONFLICT ("plugin_config_id", "key")
-        DO UPDATE SET value = posthog_pluginstorage.value::numeric + ${incrementBy}
-        RETURNING value
-        `,
-        [pluginConfigId, key, incrementBy],
-        'postgresIncrement'
-    )
-
-    return incrementResult.rows[0].value
-}
-
-export const postgresSetOnce = async (
-    db: DB,
-    pluginConfigId: PluginConfig['id'],
-    key: string,
-    value: number
-): Promise<void> => {
-    await db.postgresQuery(
-        `
-        INSERT INTO posthog_pluginstorage (plugin_config_id, key, value)
-        VALUES ($1, $2, $3)
-        ON CONFLICT ("plugin_config_id", "key")
-        DO NOTHING
-         `,
-        [pluginConfigId, key, value],
-        'postgresSetOnce'
-    )
-}
-
 export const convertDatabaseEventToPluginEvent = (
     event: Omit<Event, 'id' | 'elements' | 'elements_hash'>
 ): PluginEvent => {
@@ -172,22 +132,4 @@ export const convertDatabaseEventToPluginEvent = (
         site_url: '',
         sent_at: created_at,
     }
-}
-
-export const addPublicJobIfNotExists = async (
-    db: DB,
-    pluginId: number,
-    jobName: string,
-    jobPayloadJson: Record<string, any>
-): Promise<void> => {
-    await db.postgresQuery(
-        `
-        UPDATE posthog_plugin
-        SET public_jobs = public_jobs || $1::jsonb
-        WHERE id = $2
-        AND (SELECT (public_jobs->$3) IS NULL)
-         `,
-        [JSON.stringify({ [jobName]: jobPayloadJson }), pluginId, jobName],
-        'addPublicJobIfNotExists'
-    )
 }
